Last version

+++ solar_input.py
@@ -51,11 +51,11 @@ def parse_star_parameters(line, star):
 
     star.R = int(line.split()[1])
     star.color = line.split()[2]
-    star.m = int(line.split()[3])
-    star.x = int(line.split()[4])
-    star.y = int(line.split()[5])
-    star.Vx = int(line.split()[6])
-    star.Vy = int(line.split()[7])
+    star.m = float(line.split()[3])
+    star.x = int(float(line.split()[4]))
+    star.y = int(float(line.split()[5]))
+    star.Vx = int(float(line.split()[6]))
+    star.Vy = int(float(line.split()[7]))
     return star
 
 
@@ -76,11 +76,11 @@ def parse_planet_parameters(line, planet):
     """
     planet.R = int(line.split()[1])
     planet.color = line.split()[2]
-    planet.m = int(line.split()[3])
-    planet.x = int(line.split()[4])
-    planet.y = int(line.split()[5])
-    planet.Vx = int(line.split()[6])
-    planet.Vy = int(line.split()[7])
+    planet.m = int(float(line.split()[3]))
+    planet.x = int(float(line.split()[4]))
+    planet.y = int(float(line.split()[5]))
+    planet.Vx = int(float(line.split()[6]))
+    planet.Vy = int(float(line.split()[7]))
     return planet
